# src/main/java/com/example/gascalendar/dto/response/SuccessResponse.java
package com.example.gascalendar.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class SuccessResponse {
    private Boolean success;
    private String message;
}