# swagger.yaml
openapi: 3.0.3
info:
  title: GAS API - Goals, Actions, Steps
  description: |
    Production-ready REST API for the GAS task management and scheduling application.

    ## Features
    - JWT-based authentication
    - Kanban board task management
    - Meeting scheduling and availability
    - Admin panel for user and system management

    ## Authentication
    All endpoints except `/auth/login` and `/auth/register` require a valid JWT token.
    Include the token in the Authorization header: `Bearer <token>`

  version: 1.0.0
  contact:
    name: API Support
    email: support@gas-app.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.gas-app.com
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                summary: Admin user
                value:
                  email: admin@gas-app.com
                  password: admin123
                  rememberMe: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    success: true
                    message: Login successful
                    data:
                      user:
                        id: "1"
                        name: John Doe
                        email: admin@gas-app.com
                        role: admin
                        status: active
                        lastActive: "2026-05-13T10:30:00Z"
                        createdAt: "2025-01-01T00:00:00Z"
                      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Invalidate current session token
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve authenticated user information
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # TASKS ENDPOINTS
  # ==========================================
...
  /meetings:
    get:
      tags:
        - Meetings
      summary: Get all meetings
      description: Retrieve all scheduled meetings
      operationId: getMeetings
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
          description: Filter by meeting status
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Filter meetings from this date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: Filter meetings until this date
      responses:
        '200':
          description: Meetings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Meeting'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Meetings
      summary: Create a new meeting
      description: Schedule a new meeting with participants
      operationId: createMeeting
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingCreateRequest'
            examples:
              teamMeeting:
                summary: Team standup
                value:
                  title: Daily Standup
                  date: "2026-05-14"
                  time: "09:00"
                  participants:
                    - "1"
                    - "2"
                    - "3"
      responses:
        '201':
          description: Meeting created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Meeting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /meetings/{meetingId}:
    delete:
      tags:
        - Meetings
      summary: Cancel a meeting
      description: Cancel/delete a scheduled meeting
      operationId: deleteMeeting
      security:
        - BearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
          description: Meeting ID
      responses:
        '200':
          description: Meeting cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /meetings/{meetingId}/confirm:
    post:
      tags:
        - Meetings
      summary: Confirm or decline meeting
      description: Participant confirms, declines, or marks as pending
      operationId: confirmMeeting
      security:
        - BearerAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
          description: Meeting ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [accepted, declined, pending]
                  description: Confirmation status
                  example: accepted
      responses:
        '200':
          description: Meeting confirmation updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Meeting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /meetings/availability:
    get:
      tags:
        - Meetings
      summary: Get user availability
      description: Retrieve availability calendar for all users
      operationId: getUserAvailability
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          description: Start date for availability check
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          description: End date for availability check
      responses:
        '200':
          description: Availability retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserAvailability'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # CALENDAR ENDPOINTS
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT-based authentication. Include the token in the Authorization header:
        `Authorization: Bearer <token>`
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: admin@gas-app.com
        password:
          type: string
          format: password
          minLength: 6
          description: User password
          example: password123
        rememberMe:
          type: boolean
          default: false
          description: Keep user logged in for extended period
          example: true
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          description: User full name
          example: John Doe
        email:
          type: string
          format: email
          description: User email address
          example: john.doe@gas-app.com
        password:
          type: string
          format: password
          minLength: 6
          description: User password
          example: securePassword123
        role:
          type: string
          enum: [admin, operator, viewer]
          default: viewer
          description: User role
          example: viewer
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT access token
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "1"
        name:
          type: string
          description: User full name
          example: John Doe
        email:
          type: string
          format: email
          description: User email address
          example: john.doe@gas-app.com
        role:
          type: string
          enum: [admin, operator, viewer]
          description: User role
          example: admin
        status:
          type: string
          enum: [active, inactive]
          description: User account status
          example: active
        lastActive:
          type: string
          format: date-time
          description: Last activity timestamp
          example: "2026-05-13T10:30:00Z"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2025-01-01T00:00:00Z"
...
    Meeting:
      type: object
      properties:
        id:
          type: string
          description: Unique meeting identifier
          example: "m1"
        title:
          type: string
          description: Meeting title
          example: Team Standup
        date:
          type: string
          format: date
          description: Meeting date
          example: "2026-05-14"
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: Meeting time (HH:mm)
          example: "10:00"
        participants:
          type: array
          items:
            type: string
          description: List of participant IDs
          example: ["1", "2", "3"]
        confirmations:
          type: object
          additionalProperties:
            type: string
            enum: [pending, accepted, declined]
          description: Confirmation status per participant
          example:
            "1": accepted
            "2": pending
            "3": accepted
        status:
          type: string
          enum: [pending, confirmed, cancelled]
          description: Overall meeting status
          example: pending
        createdBy:
          type: string
          description: User ID who created the meeting
          example: "1"
        createdAt:
          type: string
          format: date-time
          example: "2026-05-13T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-05-13T10:00:00Z"
    MeetingCreateRequest:
      type: object
      required:
        - title
        - date
        - time
        - participants
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Product Review Meeting
        date:
          type: string
          format: date
          example: "2026-05-15"
        time:
          type: string
          pattern: '^\d{2}:\d{2}$'
          example: "14:00"
        participants:
          type: array
          items:
            type: string
          minItems: 1
          description: Array of participant user IDs
          example: ["1", "2", "3"]
    UserAvailability:
      type: object
      properties:
        userId:
          type: string
          example: "1"
        userName:
          type: string
          example: John Doe
        avatar:
          type: string
          description: URL to user avatar
          example: https://ui-avatars.com/api/?name=John+Doe
        availability:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Map of date to available time slots
          example:
            "2026-05-14": ["09:00", "10:00", "14:00"]
            "2026-05-15": ["11:00", "15:00"]

    # ==========================================
    # ADMIN SCHEMAS
    # ==========================================
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Human-readable error message
          example: Validation failed
        code:
          type: string
          description: Error code for programmatic handling
          example: VALIDATION_ERROR
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-specific validation errors
          example:
            email:
              - Email is required
              - Email must be valid
            password:
              - Password must be at least 6 characters
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Validation failed
            code: VALIDATION_ERROR
            errors:
              title:
                - Title is required
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication required
            code: UNAUTHORIZED
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Resource not found
            code: NOT_FOUND
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: An unexpected error occurred
            code: INTERNAL_SERVER_ERROR
security:
  - BearerAuth: []